requirements.txt
einops>=0.3.0
k-diffusion
kornia>=0.6
omegaconf>=2.1.11
torch-fidelity>=0.3.0
transformers>=4.19.2
torchmetrics>=0.6.0
-e git+https://github.com/huggingface/diffusers.git
-e git+https://github.com/openai/CLIP.git@main#egg=clip
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
